+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v1.2.1
 - F# support
 - Fix multiline display in floating windows
+- Deno brings REPL support for Javascript and TypeScript
 
 ## v1.2
 - Live mode (a @ChristianChiarulli idea and partial realisation)